feat(eslint): share helper to merge config + use of helper from eslint
