+++ packages/@o3r/eslint-config-otter/package.json
@@ -50,7 +50,7 @@
     "@typescript-eslint/eslint-plugin": "^7.14.1 || ^8.15.0",
     "@typescript-eslint/parser": "^7.14.1 || ^8.15.0",
     "@typescript-eslint/utils": "^7.14.1 || ^8.15.0",
-    "eslint": "^8.57.0 || ^9.16.0",
+    "eslint": "^8.57.0 || ^9.22.0",
     "eslint-import-resolver-node": "^0.3.4",
     "eslint-plugin-jest": "^28.8.0",
     "eslint-plugin-jsdoc": "^48.11.0 || ^49.0.0 || ^50.0.0",
