+++ packages/@o3r/eslint-config/schematics/ng-add/eslint/index.spec.ts
@@ -82,7 +82,6 @@ describe('update eslint config', () => {
     expect(tree.exists(`${libRoot}/tsconfig.eslint.json`)).toBeFalsy();
     expect(tree.readText(`${appRoot}/eslint.config.mjs`)).toContain('import shared from \'../../eslint.shared.config.mjs\'');
     expect(tree.readText(`${appRoot}/eslint.local.config.mjs`)).toContain(`${pckName}/projects`);
-    expect(tree.readText(`${appRoot}/eslint.local.config.mjs`)).toContain(`${pckName}/ignores`);
     expect(tree.readText(`${appRoot}/eslint.local.config.mjs`)).toContain('...globals.browser');
     expect(tree.readJson('angular.json')).toEqual({
       ...angularJsonContent,
@@ -117,7 +116,6 @@ describe('update eslint config', () => {
     expect(tree.exists(`${appRoot}/tsconfig.eslint.json`)).toBeFalsy();
     expect(tree.readText(`${libRoot}/eslint.config.mjs`)).toContain('import shared from \'../../eslint.shared.config.mjs\'');
     expect(tree.readText(`${libRoot}/eslint.local.config.mjs`)).toContain(`${pckName}/projects`);
-    expect(tree.readText(`${libRoot}/eslint.local.config.mjs`)).toContain(`${pckName}/ignores`);
     expect(tree.readText(`${libRoot}/eslint.local.config.mjs`)).not.toContain('...globals.browser');
     expect(tree.readJson('angular.json')).toEqual({
       ...angularJsonContent,
