+++ packages/@o3r/eslint-config/schematics/ng-add/eslint/templates/project/eslint.local.config.__extension__.template
@@ -1,14 +1,23 @@
-<% if (extension === 'mjs') { %>import { dirname } from 'node:path';
-import { fileURLToPath } from 'node:url';
+<% if (extension === 'mjs') { %>import {
+  defineConfig,
+} from 'eslint/config';
+import {
+  dirname,
+} from 'node:path';
+import {
+  fileURLToPath,
+} from 'node:url';
 <% if (isApp) {
   %>import globals from 'globals';<%
 } %>
 const __filename = fileURLToPath(import.meta.url);
 // __dirname is not defined in ES module scope
 const __dirname = dirname(__filename);
-<% } else if (isApp) { %>const globals = require('globals');
+<% } else { %>const { defineConfig } = require('eslint/config');
+<% if (isApp) { %>const globals = require('globals');
 <% } %>
-<%= extension === 'mjs' ? 'export default' : 'module.exports =' %> [
+<% } %>
+<%= extension === 'mjs' ? 'export default' : 'module.exports =' %> defineConfig([
   {
     name: '<%= packageName %>/projects',
     languageOptions: {
@@ -23,9 +32,5 @@ const __dirname = dirname(__filename);
         ...globals.browser
       }<% } %>
     }
-  },
-  {
-    name: '<%= packageName %>/ignores',
-    ignores: []
   }
-];
+]);
