+++ packages/@o3r/eslint-config/schematics/ng-add/eslint/templates/workspace/eslint.local.config.__extension__.template
@@ -1,11 +1,19 @@
-<% if (extension === 'mjs') { %>import { dirname } from 'node:path';
-import { fileURLToPath } from 'node:url';
+<% if (extension === 'mjs') { %>import {
+  defineConfig,
+} from 'eslint/config';
+import {
+  dirname,
+} from 'node:path';
+import {
+  fileURLToPath,
+} from 'node:url';
 
 const __filename = fileURLToPath(import.meta.url);
 // __dirname is not defined in ES module scope
 const __dirname = dirname(__filename);
-
-<% } %><%= codeBeforeConfig %><%= extension === 'mjs' ? 'export default' : 'module.exports =' %> [
+<% } else { %>const { defineConfig } = require('eslint/config');
+<% } %>
+<%= codeBeforeConfig %><%= extension === 'mjs' ? 'export default' : 'module.exports =' %> defineConfig([
   {
     name: '<%= packageName %>/projects',
     languageOptions: {
@@ -18,5 +26,5 @@ const __dirname = dirname(__filename);
       }
     }
   }
-].concat(<%= oldConfig %>);
+].concat(<%= oldConfig %>));
 <%= codeAfterConfig %>
