+++ tsconfig.build.json
@@ -34,6 +34,9 @@
       "@o3r/dynamic-content": ["packages/@o3r/dynamic-content/dist", "packages/@o3r/dynamic-content/src/public_api"],
       "@o3r/dynamic-content/*": ["packages/@o3r/dynamic-content/dist/*", "packages/@o3r/dynamic-content/src/*/public_api"],
       "@o3r/eslint-config": ["packages/@o3r/eslint-config"],
+      "@o3r/eslint-config/helpers": ["packages/@o3r/eslint-config/src/helpers"],
+      "@o3r/eslint-config/typescript": ["packages/@o3r/eslint-config/src/rules/typescript"],
+      "@o3r/eslint-config/template": ["packages/@o3r/eslint-config/src/rules/template"],
       "@o3r/eslint-config-otter": ["packages/@o3r/eslint-config-otter"],
       "@o3r/eslint-plugin": ["packages/@o3r/eslint-plugin/dist", "packages/@o3r/eslint-plugin/src/public_api"],
       "@o3r/extractors": ["packages/@o3r/extractors/dist", "packages/@o3r/extractors/src/public_api"],
